fix: getlaststringerror segmentfault
